break out helpers to separate script for maintainability
