+++ src/vs/workbench/api/browser/positron/mainThreadNotebookFeatures.ts
@@ -17,6 +17,7 @@ import { CellKind, CellEditType } from '../../../contrib/notebook/common/noteboo
 import { DisposableStore } from '../../../../base/common/lifecycle.js';
 import { encodeBase64 } from '../../../../base/common/buffer.js';
 import { ILogService } from '../../../../platform/log/common/log.js';
+import { isImageMimeType, isTextBasedMimeType } from '../../../contrib/positronNotebook/browser/notebookMimeUtils.js';
 
 /**
  * Main thread implementation of notebook features for extension host communication.
@@ -73,65 +74,6 @@ export class MainThreadNotebookFeatures implements MainThreadNotebookFeaturesSha
 		}
 
 		return baseDTO;
-	}
-
-	/**
-	 * Checks if a MIME type represents an image that should be base64 encoded.
-	 * @param mimeType The MIME type to check.
-	 * @returns True if the MIME type is an image type that requires base64 encoding.
-	 */
-	private isImageMimeType(mimeType: string): boolean {
-		// Image types that should be base64 encoded (excluding SVG which is text-based)
-		const imageMimeTypes = [
-			'image/png',
-			'image/jpeg',
-			'image/jpg',
-			'image/gif',
-			'image/webp',
-			'image/bmp'
-		];
-		return imageMimeTypes.includes(mimeType.toLowerCase());
-	}
-
-	/**
-	 * Checks if a MIME type represents text-based content that should be handled as plain text.
-	 * @param mimeType The MIME type to check.
-	 * @returns True if the MIME type is text-based and should be converted to string.
-	 */
-	private isTextBasedMimeType(mimeType: string): boolean {
-		// Text-based MIME types that should be converted to string
-		const textBasedMimeTypes = [
-			'text/latex',
-			'text/html',
-			'application/vnd.code.notebook.error',
-			'application/vnd.code.notebook.stdout',
-			'application/x.notebook.stdout',
-			'application/x.notebook.stream',
-			'application/vnd.code.notebook.stderr',
-			'application/x.notebook.stderr',
-			'text/plain',
-			'text/markdown',
-			'application/json'
-		];
-
-		const lowerMimeType = mimeType.toLowerCase();
-
-		// Check if it's in the known list
-		if (textBasedMimeTypes.includes(lowerMimeType)) {
-			return true;
-		}
-
-		// Check if it starts with 'text/' (covers text/xml, text/csv, etc.)
-		if (lowerMimeType.startsWith('text/')) {
-			return true;
-		}
-
-		// SVG is text-based XML
-		if (lowerMimeType === 'image/svg+xml') {
-			return true;
-		}
-
-		return false;
 	}
 
 	/**
@@ -374,7 +316,7 @@ export class MainThreadNotebookFeatures implements MainThreadNotebookFeaturesSha
 					});
 				}
 				// Handle image MIME types - base64 encode
-				else if (this.isImageMimeType(mimeType)) {
+				else if (isImageMimeType(mimeType)) {
 					const base64Data = encodeBase64(item.data);
 					outputDTOs.push({
 						mimeType: mimeType,
@@ -382,7 +324,7 @@ export class MainThreadNotebookFeatures implements MainThreadNotebookFeaturesSha
 					});
 				}
 				// Handle text-based MIME types - convert to string
-				else if (this.isTextBasedMimeType(mimeType)) {
+				else if (isTextBasedMimeType(mimeType)) {
 					outputDTOs.push({
 						mimeType: mimeType,
 						data: item.data.toString()
